requirements.txt
psycopg2
django>=1.6
south
coverage
coveralls
django-nose
django-pytest
djangorestframework
markdown
django-filter
django-oauth-toolkit
django-widget-tweaks
django-tables2
django-userena